backend/pyproject.toml
[project]
name = "krishi-saarthi"
version = "0.1.0"
description = "Backend for the Krishi Saarthi App"
authors = [
    {name = "Divyansh Kulshreshtha",email = "k.divyansh268@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
package-mode = false
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-multipart>=0.0.6",
    "requests (>=2.32.4,<3.0.0)",
    "openai>=1.0.0",
    "twilio>=9.0.0",
    # YouTube Recommender dependencies
    "firebase-admin>=6.2.0",
    "google-generativeai>=0.3.2",
    "google-api-python-client>=2.108.0",
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
